+++ users/noah/lib/state/runtime/runtime_context_internal.h
@@ -41,8 +41,8 @@ typedef struct {
 
 typedef struct {
     noah_runtime_trace_entry_t entries[NOAH_RUNTIME_TRACE_CAPACITY];
-    uint8_t                    next_index;
-    uint8_t                    count;
+    uint16_t                   next_index;
+    uint16_t                   count;
     bool                       overflowed;
 } noah_runtime_trace_state_t;
